requirements.txt
celery==5.0.5
Django==3.1.6
djangorestframework==3.12.2
flower==0.9.3
Pillow==8.1.0
psycopg2-binary==2.8.4
redis==3.4.1